fix(types): resolve mypy strict errors across src and tests
- Configure mypy to check both src/ and tests/ directories
- Add mypy_path for custom type stubs
- Add types-pyyaml dev dependency
- Create type stubs for ultralytics, transformers, and supervision (stubs/)

Source changes:
- console.py: Add TYPE_CHECKING imports for FastAPI types with documented
  type: ignore comments for unavoidable runtime fallbacks
- dance_emotion_moves.py: Remove type: ignore from Move subclasses, fix
  numpy dtype with astype(np.float64)
- headless_personality_ui.py: Add proper return type annotations for route
  handlers
- main.py: Fix Optional type for robot parameter, add assertion for
  narrowing
- moves.py: Use np.floating[Any] for FullBodyPose type alias, fix variable
  shadowing in evaluate()
- openai_realtime.py: Add cast() for untyped methods, fix tool specs type
- tools/code.py: Handle TextBlock type from anthropic response
- tools/commit_rules.py: Add explicit List[CheckResult] type annotation
- tools/core_tools.py: Fix get_concrete_subclasses parameter type, handle
  optional args_json
- tools/github_*.py: Add Dict[str, Any] annotations, handle Optional types,
  fix commit message bytes handling
- tools/move_head.py: Add runtime validation and cast for Direction literal
- tools/play_emotion.py: Rename RECORDED_MOVES to _RECORDED_MOVES with
  proper Optional type
- utils.py: Fix list return type, rename HeadTracker imports to avoid
  redefinition
- vision/processors.py: Add explicit type annotations, add unreachable
  return for mypy
- vision/yolo_head_tracker.py: Remove unnecessary type: ignore

Test changes:
- Add proper type annotations to fixtures and test functions
- Replace __builtins__["__import__"] with builtins.__import__
- Use cast(MagicMock, ...) for mock assertions
- Use object.__setattr__ for method assignment on typed objects
- Create supervision type stub for Detections class
- Fix dict type parameters (dict[str, str] instead of bare dict)

+++ src/reachy_mini_conversation_app/dance_emotion_moves.py
@@ -19,7 +19,7 @@ from reachy_mini_dances_library.dance_move import DanceMove
 logger = logging.getLogger(__name__)
 
 
-class DanceQueueMove(Move):  # type: ignore
+class DanceQueueMove(Move):
     """Wrapper for dance moves to work with the movement queue system."""
 
     def __init__(self, move_name: str):
@@ -49,11 +49,11 @@ class DanceQueueMove(Move):  # type: ignore
             # Return neutral pose on error
             from reachy_mini.utils import create_head_pose
 
-            neutral_head_pose = create_head_pose(0, 0, 0, 0, 0, 0, degrees=True)
+            neutral_head_pose = create_head_pose(0, 0, 0, 0, 0, 0, degrees=True).astype(np.float64)
             return (neutral_head_pose, np.array([0.0, 0.0], dtype=np.float64), 0.0)
 
 
-class EmotionQueueMove(Move):  # type: ignore
+class EmotionQueueMove(Move):
     """Wrapper for emotion moves to work with the movement queue system."""
 
     def __init__(self, emotion_name: str, recorded_moves: RecordedMoves):
@@ -83,11 +83,11 @@ class EmotionQueueMove(Move):  # type: ignore
             # Return neutral pose on error
             from reachy_mini.utils import create_head_pose
 
-            neutral_head_pose = create_head_pose(0, 0, 0, 0, 0, 0, degrees=True)
+            neutral_head_pose = create_head_pose(0, 0, 0, 0, 0, 0, degrees=True).astype(np.float64)
             return (neutral_head_pose, np.array([0.0, 0.0], dtype=np.float64), 0.0)
 
 
-class GotoQueueMove(Move):  # type: ignore
+class GotoQueueMove(Move):
     """Wrapper for goto moves to work with the movement queue system."""
 
     def __init__(
